backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.6.0
pydantic-settings==2.1.0
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
numpy==1.26.3
pandas==2.2.0
catsim>=0.18.0
spacy==3.7.2
openai-whisper==20231117
python-dotenv==1.0.1
httpx==0.26.0
